Add more docs and tests
